* #10, #6157 - Create shared memory management class.
* 공유메모리 관리용 인터페이스 생성
* 테스트 프로그램 작성
* CA에서 사용하는 인터페이스를 기준으로 공유메모리 클래스 작성 시작

+++ src/common/SCSelectReactor.cpp
@@ -144,7 +144,7 @@ void SCSelectReactor::set_sockethandle(fd_set& fdset, int& maxfd)
 	{
 		if (iter_s->first != SC_SOCKET_INVALID)
 		{
-			if (maxfd < iter_s->first)
+			if (maxfd < (int)iter_s->first)
 			{
 				maxfd = (int)iter_s->first;
 			}
